Rewrite tethertest to use direct invocation instead of tether.Page
The harness previously wrapped tether.Page and read assertion data from
a JSON-deserialised wire response, causing type loss — all numeric
signal values became float64. It also executed the handler twice per
Send call (once through Page, once locally for state).

Replace the HTTP round-trip with direct handler invocation: a single
dispatch call, effects captured from the testSession, HTML rendered
from the node tree. This preserves original Go types in signal
assertions (HasSignal("counter", 1) works without float64 casts),
eliminates the double execution, and removes dependencies on
encoding/json, net/http, and net/http/httptest.

Config.Render is now optional — tests that only check state and effects
no longer need a render function. An optional Config.Layout field
supports wrapping content for Render() output.

+++ tethertest/tethertest_test.go
@@ -183,10 +183,10 @@ func TestHasSignal(t *testing.T) {
 	h := newHarness()
 	h.Send("signal")
 
-	if !h.HasSignal("count", float64(0)) {
+	if !h.HasSignal("count", 0) {
 		t.Errorf("HasSignal(count, 0) = false, want true; signals = %v", h.Signals())
 	}
-	if h.HasSignal("count", float64(99)) {
+	if h.HasSignal("count", 99) {
 		t.Error("HasSignal(count, 99) = true, want false")
 	}
 	if h.HasSignal("missing", nil) {
@@ -295,9 +295,8 @@ func TestNavigateWithPath(t *testing.T) {
 }
 
 func TestMiddleware(t *testing.T) {
-	// Encode call order into state so we verify the local re-derivation
-	// path rather than capturing closure side effects from both HTTP and
-	// local paths.
+	// Encode call order into state so we verify middleware wrapping
+	// order directly.
 	outer := func(next tethertest.HandleFunc[state]) tethertest.HandleFunc[state] {
 		return func(sess tether.Session, s state, ev tether.Event) state {
 			s.Name += "A"
